- Changed Message into an interface, default impl is BaseMessage (https://issues.jboss.org/browse/JGRP-2218) - Added impls BytesMessage (default) and ObjectMessage, test suite passes 100% - Changed FRAG, FRAG2, FRAG3, COMPRESS, SYM/ASYM_ENCRYPT to fragment messages other than BytesMessages (e.g. ObjectMessages) - Changed Message.size() to return int instead of long - Moved Message.setTransientFlag(TransientFlag... flags) to setFlag(TransientFlag... flags) - Renamed Buffer -> ByteArray - SEQUENCER now handles messages other than BytesMessage correctly - Added ObjectMessageTest - StreamableTo/FromBuffer() and messageTo/FromBuffer now throw an exception - Created EmptyMessage and using it instead of ByteMessage to send messages that contain only headers but no payload - Added ObjectMessage / ObjectMessageSerializable - New class PartialDataOutputStream - Added FragmentedMessage - Added FRAG4 (using FragmentedMessage) - Added FRAG4_Test - Added NioMessage, CompositeMessage - This version has a 100% unit tests pass rate - SERIALIZE sends a BytesMessage, rather than copying the incoming message - Changed version to 5.0.0-SNAPSHOT - Moved unit tests from FRAG4_Test --> FragTest - Merged NioDirectMessage into NioMessage (no need to have separate classes) - Changed Message interface to return <T extends Message> T instead of Message (no casts needed) - Added tests to CompositeMessageTest - Added tests for CompositeMessage to FragTest - Added COMPRESS_Test - Enrypt now supports messages other than BytesMessages (test: ASYM_ENCRYPT_Test.testObjectMessage()) - Updated tutorial - Updated the manual - Message -> EmptyMessage

+++ src/org/jgroups/protocols/NoBundler.java
@@ -26,7 +26,7 @@ public class NoBundler implements Bundler {
     public void stop()  {}
 
     public void send(Message msg) throws Exception {
-        ByteArrayDataOutputStream out=new ByteArrayDataOutputStream((int)(msg.size() + 10));
+        ByteArrayDataOutputStream out=new ByteArrayDataOutputStream(msg.size() + 10);
         sendSingleMessage(msg, out);
     }
